% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detect_sparsity}
\alias{detect_sparsity}
\title{Detects whether an ARFF file is in sparse format}
\usage{
detect_sparsity(arff_data)
}
\arguments{
\item{arff_data}{Content of the data section}
}
\value{
Boolean, TRUE when the file is sparse
}
\description{
Detects whether an ARFF file is in sparse format
}

