% Generated by roxygen2 (4.0.2): do not edit by hand
\name{==.mldr}
\alias{==.mldr}
\title{Checks if two mldr objects have the same structure}
\usage{
\method{==}{mldr}(mldr1, mldr2)
}
\arguments{
\item{mldr1}{First \code{mldr} object to compare}

\item{mldr2}{Second \code{mldr} object to compare}
}
\value{
true if the two mldr objects have the same structure
}
\description{
Checks if two mldr objects have the same structure
}

