% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{genbase}
\alias{genbase}
\title{genbase}
\format{An mldr object with 662 instances, 1213 attributes and 27 labels}
\source{
S. Diplaris, G. Tsoumakas, P. Mitkas and I. Vlahavas. Protein Classification with Multiple Algorithms, Proc. 10th Panhellenic Conference on Informatics (PCI 2005), pp. 448-456, Volos, Greece, November 2005
}
\usage{
genbase
}
\description{
genbase dataset.
}
\examples{
summary(genbase)
genbase$labels
}
\keyword{datasets}

