% Generated by roxygen2 (4.0.2): do not edit by hand
\name{labelCoocurrencePlot}
\alias{labelCoocurrencePlot}
\title{Plots a graphic with co-occurrence ratio between pairs of labels}
\usage{
labelCoocurrencePlot(mld, title, labelIndices)
}
\arguments{
\item{mld}{An \code{mldr} object with the information to plot}

\item{title}{A title to show above the plot. Defaults to the name of the dataset passed as first argument}

\item{labelIndices}{Establishes the labels to show in the plot}

\item{labelCount}{Samples the labels in the dataset to show information of only \code{labelCount} of them}
}
\value{
The circular plot
}
\description{
Draws a circular plot with sectors representing each label
and links between them depicting label co-occurrences
}
\examples{
library(mldr)
yeast <- mldr("yeast") # Read "yeast.arff" and labels from "yeast.xml"
labelCoocurrencePlot(yeast) # Plots all labels
plot(yeast, type = "LC") # Same that above
plot(yeast) # Same that above
plot(yeast, title = "Yeast dataset") # Changes the title
plot(yeast, labelCount = 10) # Randomly selects 10 labels to plot
plot(yeast, labelIndices = yeast$label$index[1:10]) # Plots info of first 10 labels
}
\seealso{
\code{\link{plot.mldr}}
}

