% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{mlcr}
\alias{mlcr}
\title{Multilabel classifier constructor}
\usage{
mlcr(type, ...)
}
\arguments{
\item{type}{String identifying the algorithm to be used}
}
\description{
Creates an object of class 'mlcr'. No algorithms are provided within
  the `mldr` package, instead new packages that provide them can be installed or
  single code files can be used.
}
\examples{
classifier <- mlcr("mlknn", emotions)
}
\seealso{
\code{\link{print.mlcr}}
}

