% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mldr_from_dataframe}
\alias{mldr_from_dataframe}
\title{Generates an mldr object from a data.frame and a vector with label indices}
\usage{
mldr_from_dataframe(dataframe, labelIndices, name = NULL)
}
\description{
Generates an mldr object from a data.frame and a vector with label indices
}

