% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{mldr_roc}
\alias{mldr_roc}
\title{ROC curve}
\usage{
mldr_roc(true_labels, predicted_labels)
}
\arguments{
\item{true_labels}{Matrix of true labels, columns corresponding to labels and
rows to instances.}

\item{predicted_labels}{Matrix of predicted labels, columns corresponding to
labels and rows to instances.}
}
\value{
ROC object from pROC package
}
\description{
Calculates the ROC (Receiver Operating Characteristic) curve
 for given true labels and predicted ones. The pROC package is needed for
 this functionality.
}
