% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{mldr_to_labels}
\alias{mldr_to_labels}
\title{Label matrix of an MLD}
\usage{
mldr_to_labels(mldr)
}
\arguments{
\item{mldr}{\code{"mldr"} object.}
}
\value{
Numeric matrix of labels.
}
\description{
Extracts a matrix with the true 0-1 assignment of labels of an
 \code{"mldr"} object.
}
\seealso{
\link{Basic metrics}, \link{Averaged metrics}, \link{Ranking-based metrics}.
}
