% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mldr_transform}
\alias{mldr_transform}
\title{Transformns an MLDR into binary or multiclass datasets}
\usage{
mldr_transform(mldr, type = "BR", labels)
}
\arguments{
\item{mldr}{The mldr object to transform}

\item{type}{Indicates the type of transformation to apply. Possible types are:
"BR" Produces one or more binary datasets, each one with one label
"LP" Produces a multiclass dataset using each labelset as class label}

\item{labels}{Vector with the label indexes to include in the transformation. All labels will be used if not specified}
}
\value{
A list containing the resulting datasets
}
\description{
Transforms an \code{mldr} object into one or serveral binary or multiclass datasets, returning them as \code{data.frame} objects
}
\examples{
library(mldr)
emotionsbr <- mldr_transform(emotions, type = "BR")
}

