% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_attributes}
\alias{parse_attributes}
\title{Reads the attributes section of an ARFF file}
\usage{
parse_attributes(arff_attrs)
}
\arguments{
\item{arff_attrs}{Lines containing the attributes}
}
\value{
A vector containing, for each
 attribute, its name and its type
}
\description{
Reads the attributes section of an ARFF file
}

