% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_sparse_data}
\alias{parse_sparse_data}
\title{Builds a data.frame out of sparse ARFF data}
\usage{
parse_sparse_data(arff_data, num_attrs)
}
\arguments{
\item{arff_data}{Content of the data section}
}
\value{
data.frame containing data values
}
\description{
Builds a data.frame out of sparse ARFF data
}

