% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mldr}
\alias{plot.mldr}
\title{Generates graphic representations of an mldr object}
\usage{
\method{plot}{mldr}(mld, type = "LC", labelCount, labelIndices,
  title = NULL, ...)
}
\arguments{
\item{type}{Indicates the type of plot to produce. Possible types are:\itemize{
 \item \code{"LC"} Draws a circular plot with sectors representing each label
and links between them depicting label co-occurrences
 \item \code{"LH"} for label histogram
 }}

\item{labelCount}{Samples the labels in the dataset to show information of only \code{labelCount} of them}

\item{labelIndices}{Establishes the labels to show in the plot}

\item{title}{A title to show above the plot. Defaults to the name of the dataset passed as first argument}

\item{mldr}{The mldr object whose features are going to be drawn}
}
\description{
Generates graphic representations of an \code{mldr} object
}
\examples{
library(mldr)

# Label concurrence plot
plot(genbase, type = "LC") # Plots all labels
plot(genbase) # Same that above
plot(genbase, title = "genbase dataset") # Changes the title
plot(genbase, labelCount = 10) # Randomly selects 10 labels to plot
plot(genbase, labelIndices = genbase$label$index[1:10]) # Plots info of first 10 labels

# Label histogram plot
plot(emotions, type = "LH")
}

