% Generated by roxygen2 (4.0.2): do not edit by hand
\name{+.mldr}
\alias{+.mldr}
\title{Generates a new mldr object joining the rows
in the two mldrs given as input}
\usage{
\method{+}{mldr}(mldr1, mldr2)
}
\arguments{
\item{mldr1,}{\code{mldr2} original \code{mldr} objects}
}
\value{
a new \code{mldr} object with all rows in the two parameters
}
\description{
Generates a new mldr object joining the rows
in the two mldrs given as input
}

