% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{print.mlcr}
\alias{print.mlcr}
\title{Print method for multilabel classifiers}
\usage{
\method{print}{mlcr}(classifier)
}
\arguments{
\item{classifier}{`mlcr` object}
}
\description{
Outputs info about the specified classifier when it's trained.
}
\examples{
classifier <- mlcr("mlknn", emotions)
print(classifier)
}
\seealso{
\code{\link{mlcr}}
}

