% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.mldr}
\alias{print.mldr}
\title{Prints the mldr content}
\usage{
\method{print}{mldr}(mld)
}
\arguments{
\item{mld}{Object whose data are to be printed}
}
\description{
Prints the \code{mldr} object data, including input attributs and output labels
}
\examples{
library(mldr)

emotions
print(emotions) # Same as above
}
\seealso{
\code{\link{mldr}}, \code{\link{summary.mldr}}
}

