% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_arff}
\alias{read_arff}
\title{Contains necessary functions to read ARFF files in
different formats (sparse, nonsparse...)}
\usage{
read_arff(arff_file)
}
\arguments{
\item{arff_file}{Path to the file}
}
\value{
List containing the relation string,
 a named vector for attributes and a data.frame
 for the data section
}
\description{
Extracts all useful data from an ARFF file in
R objects
}

