% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_header}
\alias{read_header}
\title{Reads the name and Meka parameters in the header of an
ARFF file}
\usage{
read_header(arff_relation)
}
\arguments{
\item{arff_relation}{"relation" line of the ARFF file}
}
\value{
Number of labels in the dataset
}
\description{
Reads the name and Meka parameters in the header of an
ARFF file
}

