% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_xml}
\alias{read_xml}
\title{Reads the associated XML file for a ML dataset}
\usage{
read_xml(xml_file)
}
\arguments{
\item{xml_file}{Path to the XMl file}
}
\value{
A vector of strings containing the name
 of each label
}
\description{
Reads the associated XML file for a ML dataset
}

