% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrence.R
\name{remedial}
\alias{remedial}
\title{Decouples highly imbalanced labels}
\source{
F. Charte, A. J. Rivera, M. J. del Jesus, F. Herrera. "Resampling Multilabel Datasets by Decoupling Highly Imbalanced Labels". Proc. 2015 International Conference on Hybrid Artificial Intelligent Systems (HAIS 2015), pp. 489-501, Bilbao, Spain, 2015
}
\usage{
remedial(mld)
}
\arguments{
\item{mld}{\code{mldr} object with the multilabel dataset to preprocess}
}
\value{
An mldr object containing the preprocessed multilabel dataset
}
\description{
This function implements the REMEDIAL algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets,
whose aim is to decouple frequent and rare classes appearing in the same instance. For doing so, it aggregates new instances to the dataset
and edit the labels present in them.
}
\examples{

library(mldr)
\dontrun{
summary(birds)
summary(remedial(birds))
}
}
\seealso{
\code{\link{concurrenceReport}}, \code{\link{labelInteractions}}
}

