% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{roc}
\alias{roc}
\alias{roc.mldr}
\title{ROC curve}
\usage{
roc(...)

\method{roc}{mldr}(mldr, predictions, ...)
}
\arguments{
\item{...}{Additional parameters to be passed to the \code{pROC::roc} function.
See \code{\link[pROC]{roc}} for more information.}

\item{mldr}{An \code{"mldr"} object. Its labels will be extracted via
\code{\link{mldr_to_labels}}.}

\item{predictions}{Matrix of predicted labels or probabilities, columns
corresponding to labels and rows to instances.}
}
\value{
ROC object from pROC package.
}
\description{
Calculates the ROC (Receiver Operating Characteristic) curve
 for given true labels and predicted ones. The pROC package is needed for
 this functionality.
}
\seealso{
\code{\link{mldr_evaluate}} \code{\link[pROC]{roc}}
}
