% Generated by roxygen2 (4.0.2): do not edit by hand
\name{[.mldr}
\alias{[.mldr}
\title{Contains functions to filter rows (instances) and columns (attributes)}
\usage{
\method{[}{mldr}(mldrObject, rowFilter = T, ...)
}
\arguments{
\item{mldrObject}{original mldr object}

\item{rowFilter}{expression to filter the rows}
}
\value{
a new mldr object with the selected rows
}
\description{
Generates a new mldr object containing the
selected rows
}

