% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.mldr}
\alias{summary.mldr}
\title{Provides a summary of measures about the mldr}
\usage{
\method{summary}{mldr}(mld)
}
\arguments{
\item{mld}{Object whose measures are to be printed}
}
\description{
Prints a summary of the measures obtained from the \code{mldr} object
}
\examples{
library(mldr)

summary(emotions)
}
\seealso{
\code{\link{mldr}}, \code{\link{print.mldr}}
}

