% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{test}
\alias{test}
\title{Test method for multilabel classifiers}
\usage{
test(classifier, testSet)
}
\arguments{
\item{classifier}{`mlcr` object}

\item{testSet}{An `mldr` with the test instances}
}
\description{
Uses methods specific to the algorithm in order to retrieve predictions.
}
\seealso{
\code{\link{mlcr}}
}

