% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{train}
\alias{train}
\title{Training method for multilabel classifiers}
\usage{
train(classifier, trainingSet)
}
\arguments{
\item{classifier}{`mlcr` object}

\item{trainingSet}{An `mldr` with the training instances}
}
\description{
Delegates on methods specific to the algorithm to train the classifier
}
\seealso{
\code{\link{mlcr}}
}

