% Generated by roxygen2 (4.0.2): do not edit by hand
\name{==.mldr}
\alias{==.mldr}
\title{Checks if two mldr objects have the same structure}
\usage{
\method{==}{mldr}(mldr1, mldr2)
}
\arguments{
\item{mldr1,}{mldr2 mldr objects to comare}
}
\value{
true if the two mldr have the same structure
}
\description{
Checks if two mldr objects have the same structure
}

