% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/operators.R
\name{==.mldr}
\alias{==.mldr}
\title{Checks if two mldr objects have the same structure}
\usage{
\method{==}{mldr}(mldr1, mldr2)
}
\arguments{
\item{mldr1}{First \code{mldr} object to compare}

\item{mldr2}{Second \code{mldr} object to compare}
}
\value{
\code{TRUE} if the two mldr objects have the same structure, \code{FALSE} otherwise
}
\description{
Checks if two mldr objects have the same structure
}

