% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmetrics_rk.R
\name{Ranking-based metrics}
\alias{Ranking-based metrics}
\alias{average_precision}
\title{Multi-label ranking-based evaluation metrics}
\usage{
average_precision(true_labels, predictions, ...)
}
\arguments{
\item{true_labels}{Matrix of true labels, columns corresponding to labels and
rows to instances.}

\item{predictions}{Matrix of probabilities predicted by a classifier.}

\item{...}{Additional parameters to be passed to the \code{rank} function.}
}
\value{
Performance metric value
}
\description{
...
}
\details{
The \code{ties.method} parameter for the \code{rank} function may
 vary results in some edge cases.
}
