% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmetrics_ml.R
\name{Evaluation metrics}
\alias{Evaluation metrics}
\alias{hammingLoss}
\alias{subsetAccuracy}
\title{Multi-label evaluation metrics}
\usage{
hammingLoss(trueLabels, predictedLabels)

subsetAccuracy(trueLabels, predictedLabels)
}
\arguments{
\item{trueLabels}{Matrix of true labels, columns corresponding to labels and
rows to instances.}

\item{predictedLabels}{Matrix of predicted labels, columns corresponding to
labels and rows to instances.}
}
\value{
Resulting value in the range [0, 1]
}
\description{
Several evaluation metrics designed for multi-label problems.
}
\details{
As defined in the multi-label literature, Hamming Loss describes
 the average absolute distance between a predicted label and its true value.
 Subset Accuracy is the ratio of correctly predicted labelsets.
}
