% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_new.R
\name{metrics}
\alias{metrics}
\alias{accuracy}
\alias{recall}
\alias{precision}
\alias{Fmeasure}
\alias{microRecall}
\alias{microPrecision}
\alias{macroRecall}
\alias{macroPrecision}
\alias{microFmeasure}
\alias{macroFmeasure}
\title{Multi-label averaged evaluation metrics}
\usage{
accuracy(y_true, y_pred, ...)

recall(y_true, y_pred, ...)

precision(y_true, y_pred, ...)

Fmeasure(y_true, y_pred, ...)

microRecall(y_true, y_pred, ...)

microPrecision(y_true, y_pred, ...)

macroRecall(y_true, y_pred, ...)

macroPrecision(y_true, y_pred, ...)

microFmeasure(y_true, y_pred, ...)

macroFmeasure(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{True labels}

\item{y_pred}{Predicted labels}

\item{...}{Additional parameters for precision, recall and Fmeasure}
}
\value{
Resulting value in the range [0, 1]
}
\description{
Evaluation metrics based on simple metrics for the confusion
 matrix, averaged through several criteria.
}
\details{
\strong{Deciding a value when denominators are zero}

Additional parameter \code{undefinedValue}: The value to be returned when a
computation results in an undefined value due to a division by zero. Can be
a single value (e.g. NA, 0), a function with the following signature:

\code{function(tp, fp, tn, fn)}

or a string corresponding to one of the predefined strategies. These are:
\itemize{
 \item \code{"diagnose"}: This strategy performs the following decision:
 \itemize{
  \item Returns 1 if there are no true labels and none were predicted
  \item Returns 0 otherwise
 }
 This is the default strategy, and the one followed by MULAN.
 \item \code{"ignore"}: Occurrences of undefined values will be ignored when
 averaging (averages will be computed with potentially less values than
 instances/labels). Undefined values in micro-averaged metrics cannot be
 ignored (will return \code{NA}).
 \item \code{"na"}: Will return \code{NA} (with class \code{numeric}) and it
 will be propagated when averaging (averaged metrics will potentially return
 \code{NA}).
}
}
