% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mldr}
\alias{mldr}
\title{Creates an object representing a multilabel
dataset}
\usage{
mldr(filename = NULL, use_xml = TRUE, auto_extension = TRUE,
  xml_file = NULL)
}
\arguments{
\item{filename}{Name of the dataset}

\item{use_xml}{Specifies whether to use an
associated XML file to identify the labels.}

\item{auto_extension}{Specifies whether to add
the '.arff' and '.xml' extensions to the filename
where appropriate}

\item{xml_file}{Path to the XML file. If not
provided, the filename ending in ".xml" will be
assumed}
}
\description{
Creates an object representing a multilabel
dataset
}

