% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{mldr_evaluate}
\alias{mldr_evaluate}
\title{Evaluates the predictions made by a multilabel classifier}
\usage{
mldr_evaluate(mldr, predictions, threshold = 0.5)
}
\arguments{
\item{mldr}{Object of \code{mldr} type containing the instances to evaluate}

\item{predictions}{Matrix with the labels predicted for each instance in the \code{mldr} parameter. Each element
should be a value into [0,1] range}

\item{threshold}{Threshold to use to generate bipartition of labels. By default the value 0.5 is used}
}
\value{
A list with multilabel predictive performance measures. The items in the list will be \itemize{
 \item \code{Accuracy}
 \item \code{AUC}
 \item \code{AveragePrecision}
 \item \code{Coverage}
 \item \code{FMeasure}
 \item \code{HammingLoss}
 \item \code{MacroAUC}
 \item \code{MacroFMeasure}
 \item \code{MacroPrecision}
 \item \code{MacroRecall}
 \item \code{MicroAUC}
 \item \code{MicroFMeasure}
 \item \code{MicroPrecision}
 \item \code{MicroRecall}
 \item \code{OneError}
 \item \code{Precision}
 \item \code{RankingLoss}
 \item \code{Recall}
 \item \code{SubsetAccuracy}
 \item \code{ROC}
 }
 The \code{ROC} element corresponds to a \code{roc} object associated to the \code{MicroAUC} value. This object can be given as input to \code{plot} for plotting the ROC curve
 The \code{AUC}, \code{MacroAUC}, \code{MicroAUC} and \code{ROC} members will be \code{NULL} if the \code{pROC} package is not installed.
}
\description{
Taking as input an \code{mldr} object and a matrix with the predictions
given by a classifier, this function evaluates the classifier performance through
several multilabel metrics.
}
\examples{
\dontrun{
library(mldr)

# Get the true labels in emotions
predictions <- as.matrix(emotions$dataset[,emotions$labels$index])
# and introduce some noise (alternatively get the predictions from some classifier)
predictions[sample(1:593, 100), 1:6] <- sample(0:1, 600, replace = TRUE)
# then evaluate predictive performance
res <- mldr_evaluate(emotions, predictions)
str(res)
plot(res$ROC, main = "ROC curve for emotions")
}
}
\seealso{
\code{\link{mldr}}, \link{Basic metrics}, \link{Averaged metrics}, \link{Ranking-based metrics}, \code{\link{mldr_roc}}
}
