% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mldrGUI}
\alias{mldrGUI}
\title{Launchs the web-based GUI for mldr}
\usage{
mldrGUI()
}
\value{
Nothing
}
\description{
Loads in the web browser an interactive user interface built using R shiny.
}
\details{
The \pkg{mldr} package provides a basic, Shiny-based GUI to work with multilabel datasets.
You have to install the \pkg{shiny} package to be able to use this GUI.

The user interface allows working with any of the previous loaded datasets, as well as loading
new ones. The GUI is structured into the following pages:
\itemize{
  \item{\strong{Main:}}{ This page is divided into two sections.
  The one at the left can be used to choose apreviously loaded dataset,
  as well as to load datasets from files. The right part shows some basic
  statistics about the selected multilabel dataset.}
  \item{\strong{Labels:}}{ This page shows a table containing for each label its name, index, count, relative frequency
  and imbalance ratio (IRLbl). The labels can be sorted for any of these fields. The page also includes a bar plot
  of the label frequency. The range of labels in the plot can be customized.}
  \item{\strong{Labelsets:}}{ This page shows a table containing for each labelset its representation and a counter.}
  \item{\strong{Attributes:}}{ This page shows a table containing for each attribute its name, type and a summary
  of its values.}
}
}
\examples{
library(mldr)
mldrGUI()
}

