% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mldrGUI}
\alias{mldrGUI}
\title{Launchs the web-based GUI for mldr}
\usage{
mldrGUI()
}
\value{
Nothing
}
\description{
mldrGUI starts loads the web browser an interactive user interface built using R shiny.
}
\details{
Launchs the web-based GUI for mldr
}
\examples{
\dontrun{
library(mldr)
mldrGUI()
}
}

