% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mldr}
\alias{plot.mldr}
\title{Generates graphic repreentations of an mldr object}
\usage{
\method{plot}{mldr}(mld, type = "LC", labelCount, labelIndices,
  title = NULL, ...)
}
\arguments{
\item{type}{Indicates the type of plot to produce. Possible types are:
"LC" for Label Co-ocurrence circular plot}

\item{mldr}{The mldr object whose features are going to be drawn}
}
\description{
Generates graphic representations of an \code{mldr} object
}
\examples{
library(mldr)
plot(emotions, type = "LC")
}

