% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/operators.R
\name{+.mldr}
\alias{+.mldr}
\title{Generates a new mldr object joining the rows
in the two mldrs given as input}
\usage{
\method{+}{mldr}(mldr1, mldr2)
}
\arguments{
\item{mldr1}{First \code{mldr} object to join}

\item{mldr2}{Second \code{mldr} object to join}
}
\value{
a new \code{mldr} object with all rows in the two parameters
}
\description{
Generates a new mldr object joining the rows
in the two mldrs given as input
}

