% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{print.mlcr}
\alias{print.mlcr}
\title{Print method for multilabel classifiers}
\usage{
\method{print}{mlcr}(x, ...)
}
\arguments{
\item{x}{`mlcr` object}

\item{...}{Extra parameters. Currwntly ignored}
}
\description{
Outputs info about the specified classifier when it's trained.
}
\seealso{
\code{\link{mlcr}}
}

