% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{print.mldr}
\alias{print.mldr}
\title{Prints the mldr content}
\usage{
\method{print}{mldr}(x, ...)
}
\arguments{
\item{x}{Object whose data are to be printed}

\item{...}{Additional parameters to be given to print}
}
\description{
Prints the \code{mldr} object data, including input attributs and output labels
}
\examples{
library(mldr)

emotions
print(emotions) # Same as above
}
\seealso{
\code{\link{mldr}}, \code{\link{summary.mldr}}
}

