% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.mldr}
\alias{summary.mldr}
\title{Provides a summary of measures about the mldr}
\usage{
\method{summary}{mldr}(object, ...)
}
\arguments{
\item{object}{Object whose measures are to be printed}

\item{...}{Additional parameters to be given to print}
}
\description{
Prints a summary of the measures obtained from the \code{mldr} object
}
\examples{
library(mldr)

summary(emotions)
}
\seealso{
\code{\link{mldr}}, \code{\link{print.mldr}}
}

