% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_arff}
\alias{write_arff}
\title{Writes an \code{mldr} object to a file}
\usage{
write_arff(obj, filename, write.xml = FALSE)
}
\arguments{
\item{obj}{'mldr' object}

\item{filename}{base name for the files (without extension)}
}
\description{
Save the \code{mldr} content to an ARFF file and the label data to an XML file
}
\examples{
library(mldr)

write_arff(emotions, "myemotions")
}
\seealso{
\code{\link{mldr_from_dataframe}}, \code{\link{mldr}}
}

