% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_arff}
\alias{write_arff}
\title{Writes an ARFF and an XML file for
a mldr object}
\usage{
write_arff(obj, filename)
}
\arguments{
\item{obj}{'mldr' object}

\item{filename}{base name for the files (without extension)}
}
\description{
Writes an ARFF and an XML file for
a mldr object
}

